La til PdlApp2AppClient for å kalle hentPersonBolk
